Refactor socket and tag file paths
The previous implementation used a shared directory for the
nav sockets and tagfiles. This commit refactors the storage for
these into separate locations: a cache dir and a config dir.

The new default directories follow the standard Linux pattern
of using `.config` for configuration and `.cache` for runtime
state. Rather than using a command line argument for specifying
a directory, environment variables are now used, giving
greater flexibility to the end user.

The environment variables are as follows:
- `NAV_CACHE_DIR` - where the sockets are stored
- `NAV_CONFIG_DIR` - where the config is stored

The client, daemon, and associated tests have been updated.